PDFBOX-3065: deprecate int constructor, create long constructor and use it
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/2.0@1760402 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdfwriter/COSWriter.java
@@ -242,7 +242,7 @@ public class COSWriter implements ICOSVisitor, Closeable
     {
         // write to buffer instead of output
         setOutput(new ByteArrayOutputStream());
-        setStandardOutput(new COSStandardOutputStream(output, (int)inputData.length()));
+        setStandardOutput(new COSStandardOutputStream(output, inputData.length()));
 
         incrementalInput = inputData;
         incrementalOutput = outputStream;
